BUG: PI ID not conditional on PI having a first name field
The PI ID variable was only being set when the record has a
first name entry for the PI. Since the PI ID was otherwise unset,
this would cause the code to fail on records where there is a PI
field but no first name.

Now we extract the PI ID if there is a PI entry, regardless of
whether they have a first name.