Mock up config file for testing
When a dataset is opened and caching is enabled the cache directory will be saved to the config.toml file. After that every new call to PanDataSet will read the config file for the cache dir and use it if not otherwise specified.

Instead of working on the user's default config file a mockup CONFIG_PATH is created. Thus, PanDataSet will create the config file there.

Further, a switch from tempfile to the pytest built-in tmp_path is implemented.