Adapt download for different column names
The binary data column can have different column names and even names like "netCDF" and "netCDF_2" in one data frame.
We search for those names in the actual column names disregarding the case. We also exclude column names which match the possible names but have a "(" in it, as these column are usually auxiliary data describing the binary data.

In the interactive mode we now ask for the specific columns to be downloaded. Empty input is translated to all columns.

In the Harvester class the attribute column_names is renamed to columns and only the relevant parameters from the PanDataSet are passed instead of the whole data set.

_list_available_data is adjusted to still return a list of filenames even when the dataframe now has multiple columns.