Handle URLs in the binary column correctly
Some binary data sets list the download link in the binary column instead of a simple filename. Thus, we can just use this download link instead of constructing one.

Parse the url and extract the actual filename in this case as well to save the data to the correct local filename.

The test checks if the file is correctly downloaded.