Add download of whole binary data set via the ZIP link
PANGAEA offers an easier way to download all binary files of a binary data set via a ZIP link. However, a valid auth_token (called Bearer Token on the website) is needed for that.

The advantage of the ZIP link is that the tape archive optimizes the file retrieval and only one http request is needed.

 A further improvement is that pangaeapy now passes itself as the User-Agent inside the http request. This should make it easier for PANGAEA to track access to their database.

 Note that binary data sets with a URL column do not offer this ZIP download.

 If the user does not open the data set with a valid auth_token, we return a descriptive error.